+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Fix pressure measurements being forced to integer ([#476](https://github.com/kizniche/mycodo/issues/476))
  - Fix CCS811 Input measurement ([#467](https://github.com/kizniche/mycodo/issues/467))
  - Fix pigpio dependency install issue
+ - Prevent pre-output from remaining on after an Input is deactivated
 
 ### Miscellaneous